+++ tests/test_session.py
@@ -384,6 +384,21 @@ def test_render6():
         'role': 'assistant', 'parts': [{'text': 'I am fine.'}]}]
 
 
+def test_render7():
+    prompt = Prompt(message="Hello, how are you?", role=Role.USER)
+    response = Response(message="I am fine.")
+    session = Session(messages=[prompt, response], title="session1")
+    session.disable_message(0)
+    assert session.render() == 'I am fine.\n'
+    assert session.render(RenderFormat.OPENAI) == [{'content': 'I am fine.', 'role': 'assistant'}]
+    assert session.render(RenderFormat.AI_STUDIO) == [{'role': 'assistant', 'parts': [{'text': 'I am fine.'}]}]
+    session.enable_message(0)
+    session.disable_message(1)
+    assert session.render() == 'Hello, how are you?\n'
+    assert session.render(RenderFormat.OPENAI) == [{'content': 'Hello, how are you?', 'role': 'user'}]
+    assert session.render(RenderFormat.AI_STUDIO) == [{'role': 'user', 'parts': [{'text': 'Hello, how are you?'}]}]
+
+
 def test_check_render1():
     prompt = Prompt(message="Hello, how are you?", role=Role.USER)
     response = Response(message="I am fine.")